refactor: Assets 폴더를 만들어 Icons 폴더를 넣어줌
