+++ components/TodayTodoItem/index.tsx
@@ -2,8 +2,8 @@ import { useState } from "react";
 
 import useClick from "../../hooks/useClick";
 
-import bigCheck from "../../icons/bigCheck";
-import bigDelete from "../../icons/bigDelete";
+import bigCheck from "../../Assets/icons/bigCheck";
+import bigDelete from "../../Assets/icons/bigDelete";
 
 import { default as S } from "./style";
